refactor(p2): extract SubAgent into sibling page (stage 1 of 4)
SettingExperimentalSubAgentPage + 3 private helpers (SubAgentSelectRow,
SubAgentBuiltInRow, SubAgentCustomRow) moved from SettingExperimentalPage.kt
(L182-697) into new sibling SettingExperimentalSubAgentPage.kt.

Byte-equal: bodies and signatures unchanged; only the package boundary moves.
Visibility unchanged (public page composable, private helpers).

Coordinator strip: 2447 -> 1920 (-527 with import cleanup).
Imports trimmed in coordinator: 11 SubAgent-only imports removed
(DEFAULT_SUB_AGENT_*, EXTENDED_SUB_AGENT_*, SubAgent*, applyOverride,
rememberSaveable). AgentPromptConfigRepository kept (still used by
ModelCouncil page).

Baseline 6/463 unit tests preserved.
